Compose Meteosat-12 HRFI images from the band buttons
The HRFI entry in GeoSatellites.ini was FDHSI copied across, and HRFI
samples one step finer than FDHSI : its two solar channels sit on the
0.5 km reference grid and its two infrared ones on 1 km, where FDHSI uses
1 km and 2 km. Every grid constant was half what it should be, and the
channel names were the FDHSI spellings. The reader opens
/data/<spectrumlist entry>/measured, so vis_06 does not exist in an HRFI
file and the compose returned before it had read anything.

spectrumlist is the netCDF group names now - vis_06_hr, nir_22_hr,
ir_38_hr, ir_105_hr - imagewidth is 11136 and imagewidthhrv0 22272, and
coff/cfac with coffhrv/cfachrv are the 1 km and 0.5 km pair rather than
2 km and 1 km. imagewidth is the infrared full disc and imagewidthhrv0
the solar one, the same roles they play for FDHSI. The same values go
into the GeoSatellites.ini that is actually read at startup, which is not
in the repository.

Two tables of sixteen FDHSI channel names against their grid sizes are
gone, replaced by fciGridSize(). A channel's own name says which grid it
belongs to : the _hr suffix is HRFI, and within either product it is the
solar channels, vis_ and nir_, that get the finer of the two grids. The
two "== 11136 ? coffhrv : coff" tests became "== imagewidthhrv0", so the
navigation constants follow the satellite rather than a literal, and the
night buffer is sized and strided from mtg_total_number_of_columns[3]
instead of from 5568.

MET_12_HRFI has its own eGeoSatellite value rather than being aliased to
MET_12, with isMTGFCI() at the five places that need both - the three
projections and the two FCI navigation paths in formimage. The coastline
overlay is built on the 1 km grid and divided by factor when it is drawn,
so a 22272 wide image needs factor 0.5.

Three things laid down in 137394b did not work. newGeoTab tested
!(a != X || a != Y), which is false whatever a is, so every geostationary
tab was being built as a 41 column MTG tree. getGeostationarySegmentsMTG
read the FDHSI map whichever tab asked it. And AddSegmentsToList matched
HRFI with baseName().contains(), true on every iteration of the satellite
loop, so an HRFI file set the FDHSI list's image path as well ; it
matches on the satellite's own searchstring again, the way FDHSI does.

mtg_histogram had to widen from quint16 to quint32. A 0.5 km HRFI chunk
holds 12.4 million pixels and real imagery piles millions of them into a
single bin - bin 204, dark ocean, holds 4.18 million of chunk 20 at
2026-09-10 12:00 - so the bins wrapped and the cumulative LUT they feed
came out of it wrong. Over the whole disc that lost 4.98 million counts
and moved the 95 % stretch from 311/2088 to 313/2093.

btnGeoColor is off for HRFI. Four bands on two grids, of which only two
are solar, make no combination worth offering, and both places that
enable the button already read the satellite's colour flag.

Verified against /mnt/nfs_Vol4T/received/hvs-3/E3H-MTG-3/2026/09/10 at
the 12:00 slot, through a probe that links the application's own objects
and calls ComposeSegmentImagenetCDFMTGInThreadConcurrent with no GUI. All
four bands compose : vis_06_hr and nir_22_hr at 22272 square in about 17
seconds and 5.9 GB, ir_38_hr and ir_105_hr at 11136 in 9, every disc
centred and its limb symmetric. FDHSI still composes at both of its own
sizes, unchanged. Cropping the same fractional patch of Sahara out of
FDHSI vis_06 and HRFI vis_06_hr gives the same features in the same place
at twice the sampling, which is what confirms the grid mapping and not
only the size.

+++ core/segmentimage.h
@@ -260,7 +260,11 @@ public:
     quint16 mtg_total_number_of_rows[4];
     quint16 mtg_total_number_of_columns[4];
 
-    quint16 mtg_histogram[4][40][4096];
+    // A 0.5 km HRFI segment holds 12.4 million pixels and real FCI imagery piles
+    // millions of them into one bin - dark ocean, or the saturated top of the
+    // range - so the per segment bins have to be wider than the 65535 a quint16
+    // counts to. FDHSI overflowed this too, one chunk of it per segment.
+    quint32 mtg_histogram[4][40][4096];
 
 
     quint16 *ptrHRV[24];
